+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## [next]
+ - Adding method to load image bases on base64 data url.
 
 ## 0.20.0
 - Refactor game.dart classes into separate files